requirements.txt
django==1.6.3
south
django-cms==3.0
django-compressor
django-mptt
django-sekizai
django-debug-toolbar
django-reversion==1.8.0
django-axes
django-filer
cmsplugin-filer
python-memcached
mysql-python
django-model-utils
djangocms-admin-style
djangocms-text-ckeditor
django-braces
